Implement `ready()` method in Django application configuration class to activate signals
This commit implements the `ready()` method within the Django application configuration class. The `ready()` method is called by Django when the application is ready to be used. In this case, the method is used to import and activate the signals defined in the `users.signals` module. By importing the `users.signals` module within the `ready()` method, the signals defined in that module will be connected and functional when the application is ready. This allows the signals to respond to specific events or triggers within the application.