Define GenerateProfileImagePath for profile image paths in models
This commit introduces a new class called `GenerateProfileImagePath`, which is a deconstructible object used for generating profile image paths. The purpose of this class is to construct the path and name for the profile image file based on the instance and filename.

The `GenerateProfileImagePath` class has a `__call__` method that takes in the `instance` and `filename` as parameters. Inside this method, the extension of the filename is extracted, and a path is constructed using the `user.id` of the associated instance. The name of the profile image file is set as `'profile_image.ext'`.

Additionally, an instance of `GenerateProfileImagePath` named `user_profile_image_path` is created.

The `Profile` model is updated to include a `FileField` named `image` that allows uploading a file to the specified path generated by `user_profile_image_path`. The `upload_to` attribute of the `image` field is set to `user_profile_image_path`, which dynamically determines the upload path based on the `GenerateProfileImagePath` logic.

The `__str__` method of the `Profile` model is also defined to return a string representation of the profile, including the username of the associated user.

Note: The `import` statements for the necessary modules are already included in the commit.